+++ HISTORY.rst
@@ -146,6 +146,7 @@
 - Improve rendering of the details form.
 - Update rendering of the readonly multiselect widget to conform to Django's updated field class interface.
 - Add warning when using SQLite as the database backend.
+- Use Mailgun's flanker library to process the email sources.
 
 2.7.3 (2017-09-11)
 ==================